+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-# 0.18.0
+# 0.18.1
 
+## Minor Changes
+
+* Portrait extra small size has been changed from `20px` to `25px`.
+* Portrait can have a dark background.
+* Fixes issue with Portrait size when image would not render.
+* Disabled Pill's colours have been updated.
+* Individual and Business SVGs have been updated in Icon.
+
+# 0.18.0
 
 ## !! BREAKING CHANGE !!